backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
python-multipart==0.0.6
